Move ButtonWithAuthz to auth module, update parameters of button, add generic unauthorized message with examples in demo app locales
