+++ demo/public/locales/cs/translation.json
@@ -1,4 +1,5 @@
 {
 	"Welcome to React": "Vitejte v Reactu",
-	"Good day": "Dobry den"
+	"Good day": "Dobry den",
+	"You do not have rights": "Váš přístup je omezen, nemáte dostatečná oprávnění"
 }
