+++ demo/public/locales/en/translation.json
@@ -1,4 +1,5 @@
 {
 	"Welcome to React": "Hello my friend",
-	"Good day": "Have a nice day"
+	"Good day": "Have a nice day",
+	"You do not have rights": "Your access is restricted, you don't have rights to proceed"
 }
